Migrate all the code base to support nextjs-mf
