+++ package.json
@@ -4,9 +4,10 @@
   "private": true,
   "scripts": {
     "dev": "NODE_OPTIONS='--inspect' next dev --turbopack",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "build": "NEXT_PRIVATE_LOCAL_WEBPACK=true next build",
+    "start": "npx serve@latest out",
+    "lint": "next lint",
+    "deploy": "gh-pages -d out"
   },
   "dependencies": {
     "@shadcn/ui": "^0.0.4",
@@ -22,7 +23,8 @@
     "react-leaflet": "^5.0.0",
     "recharts": "^2.15.1",
     "swr": "^2.3.3",
-    "tailwind-merge": "^3.0.2"
+    "tailwind-merge": "^3.0.2",
+    "webpack": "^5.98.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
@@ -35,6 +37,7 @@
     "autoprefixer": "^10.4.21",
     "eslint": "^9",
     "eslint-config-next": "15.2.2",
+    "gh-pages": "^6.3.0",
     "postcss-import": "^16.1.0",
     "postcss-nesting": "^13.0.1",
     "typescript": "^5"
