+++ src/components/ThemeToggle.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useTheme } from './ThemeProvider'
-import { Button } from './ui/Button'
+import Button from './ui/Button'
 
 export function ThemeToggle() {
   const { theme, toggleTheme } = useTheme()
